refactor: redesign card
